requirements.txt
numpy==2.3.1
opencv_python==4.12.0.88
Pillow==11.3.0
pywinauto==0.6.9
